requirements.txt
Django==5.0.4
djangorestframework==3.15.1
stripe==9.5.0
# Add other dependencies here
